Changed many "int" to "size_t".

+++ fmesher/src/basis.cc
@@ -41,11 +41,11 @@ namespace fmesh {
   }
 
   Matrix<double> spherical_harmonics(const Matrix3<double>& S,
-				     int max_order,
+				     size_t max_order,
 				     bool rotationally_symmetric)
   {
     Matrix<double> sph(sph_basis_n(max_order,rotationally_symmetric));
-    int i, k, m;
+    size_t i, k, m;
     double GSL_res_array[max_order+1];
 
     if (rotationally_symmetric) {    
@@ -58,7 +58,7 @@ namespace fmesh {
     } else {
         double phi, scaling_sin, scaling_cos;
 
-	int Idxs2[max_order + 1];
+	size_t Idxs2[max_order + 1];
 	// 0, 2, 6, 12, 20, 30, 42, 56, 72, 90, 110, 132, 156, 182, 210, 240
 	for (k = 0; k <= max_order; k++) {
 	  Idxs2[k] = k*(k+1);
@@ -89,8 +89,8 @@ namespace fmesh {
 
 
   Matrix<double> spherical_bsplines(const Matrix3<double>& S,
-				    int n_basis,
-				    int degree,
+				    size_t n_basis,
+				    size_t degree,
 				    bool uniform_knot_angle_spacing)
   {
     Matrix<double> basis(n_basis);
@@ -98,22 +98,22 @@ namespace fmesh {
     double s,s1,s2;
     Matrix<double> control[n_basis];
     Matrix<double> control_work[degree+1];
-    int interval;
+    size_t interval;
 
-    for (int i=0; i<=degree; i++) {
+    for (size_t i=0; i<=degree; i++) {
       knots[i] = -1.0;
     }
-    for (int i=degree+1; i<n_basis; i++) {
+    for (size_t i=degree+1; i<n_basis; i++) {
       knots[i] = (double(i-degree)/double(n_basis-degree))*2.0-1.0;
       if (uniform_knot_angle_spacing) {
 	knots[i] = sin(knots[i]*M_PI/2.0);
       }
     }
-    for (int i=n_basis; i<=n_basis+degree; i++) {
+    for (size_t i=n_basis; i<=n_basis+degree; i++) {
       knots[i] = 1.0;
     }
 
-    for (int i=0; i<n_basis; i++) {
+    for (size_t i=0; i<n_basis; i++) {
       control[i] = Matrix<double>(n_basis);
       control[i](0,i) = 1.0;
     }
@@ -122,7 +122,7 @@ namespace fmesh {
     _LOG("n_basis\t" << n_basis << endl);
     _LOG("n_basis+degree+1\t" << n_basis+degree+1 << endl);
 
-    for (int coord_idx=0; coord_idx<S.rows(); coord_idx++) {
+    for (size_t coord_idx=0; coord_idx<S.rows(); coord_idx++) {
       s = S[coord_idx][2];
 
       _LOG("step 1, coord_idx\t" << coord_idx << endl);
@@ -131,21 +131,21 @@ namespace fmesh {
 	interval++;
       
       _LOG("step 2" << endl);
-      for (int i=0; i<=degree; i++)
+      for (size_t i=0; i<=degree; i++)
 	control_work[i] = control[i+interval-degree];
       
       _LOG("step 3" << endl);
-      for (int k=1; k<=degree; k++)
-	for (int i=degree; i>=k; i--) {
+      for (size_t k=1; k<=degree; k++)
+	for (size_t i=degree; i>=k; i--) {
 	  s1 = (knots[i+interval-k+1] - s)/(knots[i+interval-k+1]-knots[i+interval-degree]);
 	  s2 = 1.0-s1;
 
-	  for (int j=0; j<n_basis; j++)
+	  for (size_t j=0; j<n_basis; j++)
 	    control_work[i](0,j) = (s1 * control_work[i-1](0,j) +
 				    s2 * control_work[i](0,j));
 	}
       
-      for (int j=0; j<n_basis; j++)
+      for (size_t j=0; j<n_basis; j++)
 	basis(coord_idx,j) = control_work[degree](0,j);
     }
